合并测试文件
