--- weixin-java-mp/src/test/java/me/chanjar/weixin/mp/api/WxMpMiscAPITest.java
@@ -1,29 +1,0 @@
-package me.chanjar.weixin.mp.api;
-
-import com.google.inject.Inject;
-import me.chanjar.weixin.common.error.WxErrorException;
-import me.chanjar.weixin.mp.api.test.ApiTestModule;
-import org.testng.*;
-import org.testng.annotations.*;
-
-import java.util.Arrays;
-
-/**
- * @author chanjarster
- */
-@Test(groups = "miscAPI")
-@Guice(modules = ApiTestModule.class)
-public class WxMpMiscAPITest {
-
-  @Inject
-  protected WxMpService wxService;
-
-  @Test
-  public void testGetCallbackIP() throws WxErrorException {
-    String[] ipArray = this.wxService.getCallbackIP();
-    System.out.println(Arrays.toString(ipArray));
-    Assert.assertNotNull(ipArray);
-    Assert.assertNotEquals(ipArray.length, 0);
-  }
-
-}
